backend_api/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
websockets
redis
pydantic-settings
python-dotenv